Remove usage of toFile in path matchers
