+++ gradle/libs.versions.toml
@@ -8,6 +8,7 @@ json-path = "2.9.0"
 junit4 = "4.13.2"
 junit-platform = "1.11.4"
 junit-jupiter = "5.13.4"
+jimfs = "1.3.0"
 kaml = "0.85.0"
 kotlin = "2.2.0"
 kotlinx-coroutines = "1.10.2"
@@ -65,6 +66,7 @@ diffutils = { module = "io.github.java-diff-utils:java-diff-utils", version.ref 
 jayway-json-path = { module = "com.jayway.jsonpath:json-path", version.ref = "json-path" }
 jdom2 = { module = "org.jdom:jdom2", version.ref = "jdom2" }
 jetbrainsAnnotations = { module = "org.jetbrains:annotations", version = "26.0.2" }
+jimfs = { module = "com.google.jimfs:jimfs", version.ref = "jimfs" }
 junit4 = { module = "junit:junit", version.ref = "junit4" }
 kaml = { module = "com.charleskorn.kaml:kaml", version.ref = "kaml" }
 mockk = { module = "io.mockk:mockk", version.ref = "mockk" }
